+++ pkg/ds/singly_linked_list_test.go
@@ -10,7 +10,7 @@ func TestAppend(t *testing.T) {
 	list := NewList[int]()
 	list.Append(1).Append(2)
 
-	assert.Equal(t, uint(2), list.Len())
+	assert.Equal(t, 2, list.Len())
 	assert.Equal(t, []int{1, 2}, list.ToSlice())
 }
 
@@ -18,7 +18,7 @@ func TestPrepend(t *testing.T) {
 	list := NewList[int]()
 	list.Prepend(2).Prepend(1)
 
-	assert.Equal(t, uint(2), list.Len())
+	assert.Equal(t, 2, list.Len())
 	assert.Equal(t, []int{1, 2}, list.ToSlice())
 }
 
@@ -60,7 +60,7 @@ func TestRemoveFirst(t *testing.T) {
 	val, succ := list.RemoveFirst()
 	assert.Equal(t, true, succ)
 	assert.Equal(t, 3, val)
-	assert.Equal(t, uint(2), list.Len())
+	assert.Equal(t, 2, list.Len())
 	assert.Equal(t, []int{1, 2}, list.ToSlice())
 }
 
@@ -68,11 +68,11 @@ func TestDelete(t *testing.T) {
 	list := NewList[int]()
 	list.Append(3).Append(2).Append(1).Append(2)
 
-	assert.Equal(t, uint(4), list.Len())
+	assert.Equal(t, 4, list.Len())
 	node, _ := findFirst(list, 2)
 	succ := list.Delete(node)
 	assert.True(t, succ)
-	assert.Equal(t, uint(3), list.Len())
+	assert.Equal(t, 3, list.Len())
 	assert.Equal(t, []int{3, 1, 2}, list.ToSlice())
 }
