+++ pkg/ds/stack_test.go
@@ -16,5 +16,5 @@ func TestPush(t *testing.T) {
 		assert.Equal(t, i, val)
 	}
 
-	assert.Equal(t, uint(0), stack.Len())
+	assert.Equal(t, 0, stack.Len())
 }
